add generate script, fix load checkpoint on train.py

+++ factor_vae/train.py
@@ -8,8 +8,7 @@ from torch.utils.data import Dataset, DataLoader
 import pytorch_lightning as pl
 import omegaconf
 import os
-from factor_vae.utils.paths import CODE_MODEL
-
+from factor_vae.utils.paths import CODE_MODEL, ROOT
 
 @hydra.main(pkg_resources.resource_filename("factor_vae", 'config'), 'train.yaml')
 def train(config: DictConfig):
@@ -17,25 +16,29 @@ def train(config: DictConfig):
 
     pl.seed_everything(config.seed)
     if config.ckpt is not None:
-        ckpt = Path(config.ckpt)
+        ckpt = ROOT / Path(config.ckpt)
         assert ckpt.exists() and ckpt.isdir(), f"Checkpoint {ckpt} does not exist or is not a directory"
         config = OmegaConf.load(ckpt / 'config.yaml')
         os.chdir(ckpt.abspath())
+        print('loaded original config from', ckpt / 'config.yaml')
     with open('config.yaml', 'w') as f:
         omegaconf.OmegaConf.save(config, f, resolve=True)
     model: pl.LightningModule = hydra.utils.instantiate(config.model)
     train_dataset: Dataset = hydra.utils.instantiate(config.dataset.train)
     val_dataset: Dataset = hydra.utils.instantiate(config.dataset.val)
     config.batch_size = config.batch_size * 2  # each training step requires two batches as specified in the paper
+    print('double batch size to value ', config.batch_size)
 
-    CODE_MODEL.copytree('model')  # copy source code of model under experiment directory
+    if not Path('model').exists():
+        CODE_MODEL.copytree('model')  # copy source code of model under experiment directory
 
     pin_memory = 'gpu' in config.accelerator
     train_dl = DataLoader(train_dataset, batch_size=config.batch_size, pin_memory=pin_memory)
     val_dl = DataLoader(val_dataset, batch_size=config.batch_size, pin_memory=pin_memory)
     ckpt_callback = ModelCheckpoint('./', 'best',
                                     monitor='valid/loss_vae_epoch',
-                                    auto_insert_metric_name=False, save_last=True)
+                                    auto_insert_metric_name=False, save_last=True,
+                                    every_n_train_steps=config.model.log_freq)
     print_stats_image(train_dataset)
 
     callbacks = [ckpt_callback]
@@ -44,7 +47,8 @@ def train(config: DictConfig):
     trainer = pl.Trainer(callbacks=callbacks, accelerator=config.accelerator, devices=config.devices,
                          gradient_clip_val=config.gradient_clip_val,
                          gradient_clip_algorithm=config.gradient_clip_algorithm,
-                         resume_from_checkpoint=ckpt, max_steps=config.max_steps)
+                         resume_from_checkpoint=ckpt / 'best.ckpt', max_steps=config.max_steps,
+                         max_epochs=config.max_epochs)
     print(model)
     trainer.fit(model, train_dl, val_dl)
